+++ src/packages/masonry/types.ts
@@ -3,3 +3,12 @@ export interface PositionItem {
   height?: number;
   top?: number;
 }
+
+export interface MasonryItem extends Object {
+  _masonryIndex?: number;
+}
+
+export interface SectionRecord {
+  head?: number;
+  tail?: number;
+}